Fix format specifiers and usage of const char* arguments/variables.
Affected files:

* comsat/comsat.c
* examples/header.c
* examples/mailcap.c
* examples/pop3client.c
* examples/url-parse.c
* imap4d/fetch.c
* imap4d/id.c
* imap4d/list.c
* imap4d/select.c
* imap4d/status.c
* imap4d/store.c
* imap4d/sync.c
* imap4d/util.c
* include/mailutils/sieve.h
* libmu_argp/common.c
* libmu_argp/muinit.c
* libmu_sieve/actions.c
* libmu_sieve/extensions/pipe.c
* libmu_sieve/extensions/vacation.c
* libmu_sieve/prog.c
* libmu_sieve/require.c
* libmu_sieve/runtime.c
* libmu_sieve/sieve-priv.h
* libmu_sieve/sieve.y
* libproto/imap/folder.c
* libproto/imap/mbox.c
* libproto/mailer/sendmail.c
* libproto/mbox/mbox.c
* libproto/pop/pop3_list.c
* libproto/pop/pop3_stat.c
* maidag/lmtp.c
* maidag/sieve.c
* mail/copy.c
* mail/decode.c
* mail/envelope.c
* mail/eq.c
* mail/escape.c
* mail/from.c
* mail/mail.h
* mail/mailline.c
* mail/mailvar.c
* mail/size.c
* mail/util.c
* mail/write.c
* mailbox/cfg_format.c
* mailbox/folder.c
* mailbox/parse822.c
* mailbox/system.c
* messages/messages.c
* mh/folder.c
* mh/forw.c
* mh/inc.c
* mh/mh.h
* mh/mh_alias.l
* mh/mh_fmtgram.y
* mh/mh_getopt.h
* mh/mh_init.c
* mh/mh_msgset.c
* mh/mh_whatnow.c
* mh/mhn.c
* mh/pick.y
* mimeview/mimetypes.y
* python/libmu_py/debug.c
* sieve/sieve.c

+++ examples/url-parse.c
@@ -54,7 +54,7 @@ print_fvpairs (mu_url_t url)
   if (fvc == 0)
     return;
   for (i = 0; i < fvc; i++)
-    printf ("\tparam[%d] <%s>\n", i, fvp[i]);
+    printf ("\tparam[%lu] <%s>\n", (unsigned long) i, fvp[i]);
 }
 
 static void
@@ -71,7 +71,7 @@ print_query (mu_url_t url)
   if (qargc == 0)
     return;
   for (i = 0; i < qargc; i++)
-    printf ("\tquery[%d] <%s>\n", i, qargv[i]);
+    printf ("\tquery[%lu] <%s>\n", (unsigned long) i, qargv[i]);
 }
 
 int
