+++ maidag/sieve.c
@@ -57,7 +57,8 @@ _sieve_action_log (void *user_name,
     {
       size_t uid = 0;
       mu_message_get_uid (msg, &uid);
-      mu_diag_printf (MU_DIAG_NOTICE, _("%s on msg uid %d"), action, uid);
+      mu_diag_printf (MU_DIAG_NOTICE, _("%s on msg uid %lu"), action,
+		      (unsigned long) uid);
     }
   
   if (fmt && strlen (fmt))
